p012 [Refactor] レンタル料の型をintではなく正しいdoubleに修正する
メソッド抽出時のバグを修正した。こういうポカミスを防ぐために、
やはり振る舞いを確認するためのテストは不可欠なんだと感じた。

IDEの補助を受けながらリファクタリングしていけば、こういった型の
不一致などのミスは防ぎやすいかもしれない。

+++ src/main/java/refactoring/chapter01/Customer.java
@@ -46,8 +46,8 @@ public class Customer {
         return result;
     }
 
-    private int amountFor(Rental each) {
-        int thisAmount = 0;
+    private double amountFor(Rental each) {
+        double thisAmount = 0;
 
         // 1行ごとに金額を計算
         switch (each.getMovie().getPriceCode()) {
